+++ src/resolvers/Channels/typeDefs.ts
@@ -2,6 +2,7 @@ import { gql } from "apollo-server"
 
 export const channelTypeDef = gql`
   interface ChannelAttr {
+    teamName: String
     channelName: String
     description: String
   }
@@ -10,15 +11,25 @@ export const channelTypeDef = gql`
     SK: String
     createdAt: Date
     updatedAt: Date
+    teamName: String
     channelName: String
     description: String
   }
 
   extend type Mutation {
+    """
+    deprecated
+    """
     createChannel(channelName: String!, description: String): Channel @auth
+    createTeamChannel(
+      teamName: String!
+      channelName: String!
+      description: String
+    ): Channel @auth
   }
 
   extend type Query {
     queryChannels(filter: String): [Channel]
+    queryTeamChannels(teamName: String!): [Channel]
   }
 `