+++ src/resolvers/Teams/Queries/index.ts
@@ -1,0 +1,2 @@
+export { getTeam } from "./getTeam"
+export { queryTeams } from "./queryTeams"