+++ src/resolvers/Votes/Queries/index.ts
@@ -1,2 +1,0 @@
-export { queryVotesByEmail } from "./queryVotesByEmail"
-export { queryVotesByProduct } from "./queryVotesByProduct"